Allow getting encoding from multivalue headers
This commit enables determing the right encoding for headers like:
`application/json, */*`.
Some applications might send headers as the above which are legit
according to RFC 26116 section 4.2
(https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2)
The previous version didn't take that into account and was returning an
error in the determine encoding function instead of the json which would
be expected in the above example.

In this commit we split the headers based on `,` and check all values to
see if any value matches our needs.

Signed-off-by: Andreas Strikos <astrikos@gmail.com>