chore: remove unnecessary nolint directives in SQS test
Remove two unnecessary nolint directives from component/sqs/message_test.go
that were suppressing warnings for getQueueUrlWithContextErr field and
GetQueueUrl method. Both are actively used in the test code and no longer
need to be suppressed.

These nolints were likely added when the code was first introduced but are
no longer necessary as the field and method are properly utilized.